lists are partially working
